pyproject.toml
[tool.poetry]
name = "dynamo"
version = "0.2.0"
description = "Quantum entanglement!"
authors = ["trumully <truman.mulholland@gmail.com>"]
readme = "README.md"
license = "LICENSE"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = { git = "https://github.com/mikeshardmind/discord.py.git", branch = "salamander-reloaded" }
click = "^8.1.7"
toml = "^0.10.2"
python-dateutil = "^2.9.0.post0"
platformdirs = "^4.2.2"
xxhash = "^3.5.0"
msgspec = "^0.18.6"
base2048 = "^0.1.3"
numpy = "^2.1.0"
pillow = "^10.4.0"
truststore = "^0.9.2"
types-toml = "^0.10.8.20240310"
types-python-dateutil = "^2.9.0.20240906"
rapidfuzz = "^3.9.7"
pygit2 = "^1.15.1"

[tool.poetry.scripts]
dynamo = "dynamo.launcher:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pytest = "^8.3.2"
hypothesis = "^6.112.0"
pytest-asyncio = "^0.24.0"
pytest-xdist = { extras = ["psutil"], version = "^3.6.1" }

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "E", "I", "UP", "YTT", "ANN", "S", "BLE", "B", "A", "COM", "C4", "DTZ", "EM", "ISC", "G", "INP", "PIE", "T20", "Q003", "RSE", "RET", "SIM", "TID", "PTH", "ERA", "PD", "PLC", "PLE", "PLR", "PLW", "TRY", "NPY", "RUF", "ASYNC"]
ignore = ["RUF001", "G002", "S101", "PLR2004", "PLC0105", "SIM105", "C90", "ANN101", "ANN202", "ANN204", "ANN401", "PLR0912", "PLR0913", "PLR0915", "UP007", "COM812", "ISC001", "B008", "ASYNC109"]
unfixable = ["ERA"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
